+++ src/test/kotlin/io/libp2p/pubsub/PubsubRouterTest.kt
@@ -2,8 +2,6 @@ package io.libp2p.pubsub
 
 import io.libp2p.etc.types.toBytesBigEndian
 import io.libp2p.etc.types.toProtobuf
-import io.libp2p.pubsub.flood.FloodRouter
-import io.libp2p.pubsub.gossip.GossipRouter
 import io.libp2p.tools.TestChannel.TestConnection
 import io.netty.handler.logging.LogLevel
 import io.netty.util.ResourceLeakDetector
@@ -14,8 +12,9 @@ import java.time.Duration
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.TimeUnit
 
-class PubsubRouterTest {
+typealias RouterCtor = () -> PubsubRouterDebug
 
+abstract class PubsubRouterTest(val router: RouterCtor) {
     init {
         ResourceLeakDetector.setLevel(ResourceLeakDetector.Level.PARANOID)
     }
@@ -28,11 +27,11 @@ class PubsubRouterTest {
             .build()
 
     @Test
-    fun test1_Fanout() {
+    fun Fanout() {
         val fuzz = DeterministicFuzz()
 
-        val router1 = fuzz.createTestRouter(GossipRouter())
-        val router2 = fuzz.createTestRouter(GossipRouter())
+        val router1 = fuzz.createTestRouter(router())
+        val router2 = fuzz.createTestRouter(router())
         router2.router.subscribe("topic1")
 
         router1.connectSemiDuplex(router2, LogLevel.ERROR, LogLevel.ERROR)
@@ -52,17 +51,12 @@ class PubsubRouterTest {
     }
 
     @Test
-    fun test2() {
-        scenario2 { FloodRouter() }
-        scenario2 { GossipRouter() }
-    }
-
-    fun scenario2(routerFactory: () -> PubsubRouterDebug) {
+    fun scenario2() {
         val fuzz = DeterministicFuzz()
 
-        val router1 = fuzz.createTestRouter(routerFactory())
-        val router2 = fuzz.createTestRouter(routerFactory())
-        val router3 = fuzz.createTestRouter(routerFactory())
+        val router1 = fuzz.createTestRouter(router())
+        val router2 = fuzz.createTestRouter(router())
+        val router3 = fuzz.createTestRouter(router())
 
         val conn_1_2 = router1.connectSemiDuplex(router2, pubsubLogs = LogLevel.ERROR)
         val conn_2_3 = router2.connectSemiDuplex(router3, pubsubLogs = LogLevel.ERROR)
@@ -115,21 +109,17 @@ class PubsubRouterTest {
         conn_1_2.disconnect()
     }
 
+    // scenario3_StarTopology { GossipRouter().withDConstants(3, 3, 100) }
     @Test
-    fun test3() {
-        scenario3_StarTopology { FloodRouter() }
-        scenario3_StarTopology { GossipRouter().withDConstants(3, 3, 100) }
-    }
-
-    fun scenario3_StarTopology(routerFactory: () -> PubsubRouterDebug) {
+    fun StarTopology() {
         val fuzz = DeterministicFuzz()
 
         val allRouters = mutableListOf<TestRouter>()
 
-        val routerCenter = fuzz.createTestRouter(routerFactory())
+        val routerCenter = fuzz.createTestRouter(router())
         allRouters += routerCenter
         for (i in 1..20) {
-            val routerEnd = fuzz.createTestRouter(routerFactory())
+            val routerEnd = fuzz.createTestRouter(router())
             allRouters += routerEnd
             routerEnd.connectSemiDuplex(routerCenter)
         }
@@ -154,22 +144,16 @@ class PubsubRouterTest {
     }
 
     @Test
-    fun test4() {
-        println("WheelTopology  FloodRouter:")
-        scenario3_WheelTopology { FloodRouter() }
-        println("WheelTopology  GossipRouter:")
-        scenario3_WheelTopology { GossipRouter() }
-    }
-    fun scenario3_WheelTopology(routerFactory: () -> PubsubRouterDebug) {
+    fun WheelTopology() {
         val fuzz = DeterministicFuzz()
 
         val allRouters = mutableListOf<TestRouter>()
         val allConnections = mutableListOf<TestConnection>()
 
-        val routerCenter = fuzz.createTestRouter(routerFactory())
+        val routerCenter = fuzz.createTestRouter(router())
         allRouters += routerCenter
         for (i in 1..20) {
-            val routerEnd = fuzz.createTestRouter(routerFactory())
+            val routerEnd = fuzz.createTestRouter(router())
             allRouters += routerEnd
             allConnections += routerEnd.connectSemiDuplex(routerCenter)
         }
@@ -215,18 +199,11 @@ class PubsubRouterTest {
     }
 
     @Test
-    fun test5() {
-        println("10NeighborsTopology  FloodRouter:")
-        scenario4_10NeighborsTopology { FloodRouter() }
-        println("10NeighborsTopology  GossipRouter:")
-        for (d in 3..6) {
-            for (seed in 0..10) {
-                print("D=$d, seed=$seed  ")
-                scenario4_10NeighborsTopology(seed) { GossipRouter().withDConstants(d, d, d) }
-            }
-        }
+    open fun TenNeighborsTopology() {
+        doTenNeighborsTopology()
     }
-    fun scenario4_10NeighborsTopology(randomSeed: Int = 0, routerFactory: () -> PubsubRouterDebug) {
+
+    fun doTenNeighborsTopology(randomSeed: Int = 0, routerFactory: RouterCtor = router) {
         val fuzz = DeterministicFuzz().also {
             it.randomSeed = randomSeed.toLong()
         }
@@ -306,72 +283,16 @@ class PubsubRouterTest {
     }
 
     @Test
-    fun testIHaveIWant() {
+    fun PublishFuture() {
         val fuzz = DeterministicFuzz()
 
-        val allRouters = mutableListOf<TestRouter>()
-
-        val otherCount = 5
-        for (i in 1..otherCount) {
-            val r = GossipRouter().withDConstants(1, 0)
-            val routerEnd = fuzz.createTestRouter(r)
-            (routerEnd.router as GossipRouter).heartbeat // init heartbeat with current time
-            allRouters += routerEnd
-        }
-
-        // make routerCenter heartbeat trigger last to drop extra peers from the mesh
-        // this is to test ihave/iwant
-        fuzz.timeController.addTime(Duration.ofMillis(1))
-
-        val r = GossipRouter().withDConstants(3, 3, 3, 1000)
-        val routerCenter = fuzz.createTestRouter(r)
-        allRouters.add(0, routerCenter)
-
-        for (i in 1..otherCount) {
-            allRouters[i].connectSemiDuplex(routerCenter, pubsubLogs = LogLevel.ERROR)
-        }
-
-        allRouters.forEach { it.router.subscribe("topic1") }
-
-        // heartbeat for all
-        fuzz.timeController.addTime(Duration.ofSeconds(1))
-
-        val msg1 = newMessage("topic1", 0L, "Hello".toByteArray())
-        routerCenter.router.publish(msg1)
-
-        Assertions.assertTrue(routerCenter.inboundMessages.isEmpty())
-
-        val receiveRouters = allRouters - routerCenter
-
-        val msgCount1 = receiveRouters.sumBy { it.inboundMessages.size }
-        println("Messages received on first turn: $msgCount1")
-
-        // The message shouldn't be broadcasted to all peers (mesh size is limited to 3)
-        Assertions.assertNotEquals(receiveRouters.size, msgCount1)
-        receiveRouters.forEach { it.inboundMessages.clear() }
-
-        // heartbeat where ihave/iwant should be used to deliver to all peers
-        fuzz.timeController.addTime(Duration.ofSeconds(1))
-
-        val msgCount2 = receiveRouters.sumBy { it.inboundMessages.size }
-        println("Messages received on second turn: $msgCount2")
-
-        // no all peers should receive the message
-        Assertions.assertEquals(receiveRouters.size, msgCount1 + msgCount2)
-        receiveRouters.forEach { it.inboundMessages.clear() }
-    }
-
-    @Test
-    fun test_PublishFuture() {
-        val fuzz = DeterministicFuzz()
-
-        val router1 = fuzz.createTestRouter(GossipRouter())
+        val router1 = fuzz.createTestRouter(router())
 
         val msg0 = newMessage("topic1", 0L, "Hello".toByteArray())
         val publishFut0 = router1.router.publish(msg0)
         Assertions.assertThrows(ExecutionException::class.java, { publishFut0.get() })
 
-        val router2 = fuzz.createTestRouter(GossipRouter())
+        val router2 = fuzz.createTestRouter(router())
         router2.router.subscribe("topic1")
 
         router1.connectSemiDuplex(router2, LogLevel.ERROR, LogLevel.ERROR)
