+++ src/test/kotlin/io/libp2p/pubsub/flood/FloodPubsubRouterTest.kt
@@ -1,0 +1,5 @@
+package io.libp2p.pubsub.flood
+
+import io.libp2p.pubsub.PubsubRouterTest
+
+class FloodPubsubRouterTest : PubsubRouterTest(::FloodRouter)
